Fix more cross-browser issues: Firefox edition
parts of elements rotated behind the plane of the parent containing
element were not clickable in Firefox, as the click event was going to
the parent element instead of the child

also, this activates the #opened-block div only when it's needed. I
originally thought this was the issue. It probably wasn't, but it seems
cleaner than the `pointer-events: none` approach, anyhow.